[postgres] move `PopulateColumns`
