+++ lib/postgres/columns.go
@@ -1,29 +1,12 @@
 package postgres
 
 import (
-	"database/sql"
 	"fmt"
 	"log/slog"
 
 	"github.com/artie-labs/reader/lib/postgres/schema"
 	"github.com/jackc/pgx/v5"
 )
-
-func (t *Table) PopulateColumns(db *sql.DB) error {
-	cols, err := schema.DescribeTable(db, t.Schema, t.Name)
-	if err != nil {
-		return fmt.Errorf("failed to describe table %s.%s: %w", t.Schema, t.Name, err)
-	}
-
-	for _, col := range cols {
-		t.Fields.AddField(col.Name, col.Type, col.Opts)
-		// Add to original columns before mutation
-		t.OriginalColumns = append(t.OriginalColumns, col.Name)
-		t.ColumnsCastedForScanning = append(t.ColumnsCastedForScanning, castColumn(col.Name, col.Type))
-	}
-
-	return t.FindStartAndEndPrimaryKeys(db)
-}
 
 // castColumn will take a colName and return the escaped version of what we should be using to call Postgres.
 func castColumn(rawColName string, colKind schema.DataType) string {
